Auto stash before rebase of "refs/heads/testoffcode"
